+++ magda-web-client/src/Routes.tsx
@@ -6,6 +6,8 @@ import AccountPage from "Components/Account/AccountPage";
 import AccountLoginPage from "Components/Account/AccountLoginPage";
 import AccountSignInRedirectPage from "Components/Account/AccountSignInRedirectPage";
 import DatasetsSearchPage from "Components/Dataset/DatasetsSearchPage";
+import DatasetsSearchPageDraftOnly from "Components/Dataset/DatasetsSearchPageDraftOnly";
+import DatasetsSearchPagePublishedOnly from "Components/Dataset/DatasetsSearchPagePublishedOnly";
 import DatasetPage from "Components/Dataset/DatasetPage";
 import DatasetSuggestPage from "Components/Dataset/DatasetSuggestPage";
 
@@ -171,6 +173,16 @@ const Routes = () => {
             <Route
                 exact
                 path="/search"
+                component={withHeader(DatasetsSearchPagePublishedOnly, true)}
+            />
+            <Route
+                exact
+                path="/drafts"
+                component={withHeader(DatasetsSearchPageDraftOnly, true)}
+            />
+            <Route
+                exact
+                path="/all-datasets"
                 component={withHeader(DatasetsSearchPage, true)}
             />
             <Route