+++ CHANGES.md
@@ -37,6 +37,14 @@ UI:
 -   Fixed web-client code loaded & run twice
 -   Added basic spatial preview to add dataset
 
+Access Control:
+
+-   Introduced Role & Permission Structure
+-   Recognise unauthenticated users as anonymous users role
+-   Introduced Open Policy Agent as policy evaluation engine
+-   Search API will return datasets based on user's Roles & Permissions
+-   Users with access to draft datasets can see a new `drafts` tab
+
 ## 0.0.55
 
 UI: